Update Knockout-WinJS bridge to support WinJS 4.1
This commit updates the knockout-winjs project to support WinJS 4.1. Obsolete
controls are removed and new controls and any associated custom property
processors are added. Fixed binding list watch and item template watch. The
bridge now allows WinJS control properties to be bound to observables and changes
to them will reflect in the observable. For example, a string observable bound
to queryText option in AutoSuggestBox will be updated when the query changes.

Fixes #1 Update for WinJS 4.0